+++ config/settings/base.py
@@ -251,7 +251,7 @@ DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 # معلومات التطبيق
 # ---------------------------------------------------------------------------
 APP_NAME = "نظام إدارة بيانات ومرضى العيادة"
-APP_VERSION = "1.5.6"
+APP_VERSION = "1.5.7"
 DB_SCHEMA_VERSION = "3"
 FACILITY_NAME = config("FACILITY_NAME", default="عيادة الموصل الخيرية")
 UPDATE_MANIFEST_URL = config("UPDATE_MANIFEST_URL", default="")